Separate view model into LibraryContainerViewModel
