+++ src/DynamoCoreWpf/ViewModels/Core/DynamoViewModel.cs
@@ -48,17 +48,6 @@ namespace Dynamo.ViewModels
 
     public partial class DynamoViewModel : ViewModelBase, IDynamoViewModel
     {
-        public void OnLibraryContainerClicked(string id)
-        {
-            var nodeSearchElement = SearchViewModel.Model.SearchEntries
-                .Where(x => x.CreationName == id).FirstOrDefault();
-            if (nodeSearchElement != null)
-            {
-                var nodeModel = nodeSearchElement.CreateNode();
-                SearchViewModel.OnSearchElementClicked(nodeModel, new Point(0, 0));
-            }
-        }
-
         public int ScaleFactorLog
         {
             get
@@ -427,8 +416,6 @@ namespace Dynamo.ViewModels
 
         public SearchViewModel SearchViewModel { get; private set; }
 
-        public LibraryViewModel LibraryViewModel { get; private set; }
-
         public PackageManagerClientViewModel PackageManagerClientViewModel { get; private set; }
 
         /// <summary>
@@ -528,7 +515,6 @@ namespace Dynamo.ViewModels
             this.WatchHandler = startConfiguration.WatchHandler;
             var pmExtension = model.GetPackageManagerExtension();
             this.PackageManagerClientViewModel = new PackageManagerClientViewModel(this, pmExtension.PackageManagerClient);
-            this.LibraryViewModel = new LibraryViewModel(this);
             this.SearchViewModel = new SearchViewModel(this);
 
             // Start page should not show up during test mode.
@@ -2440,18 +2426,5 @@ namespace Dynamo.ViewModels
         }
 
         #endregion
-    }
-
-    public class LibraryViewModel
-    {
-        private readonly Search.NodeSearchModel nodeSearchModel;
-        private readonly DynamoViewModel dynamoViewModel;
-
-        internal LibraryViewModel(DynamoViewModel dynamoViewModel)
-        {
-            nodeSearchModel = dynamoViewModel.Model.SearchModel;
-            this.dynamoViewModel = dynamoViewModel;
-
-        }
     }
 }
