+++ src/DynamoCoreWpf/Views/Core/DynamoView.xaml.cs
@@ -70,7 +70,7 @@ namespace Dynamo.Controls
         private readonly LoginService loginService;
         internal ViewExtensionManager viewExtensionManager = new ViewExtensionManager();
         private ShortcutToolbar shortcutBar;
-        private ILibraryContainer libraryContainer;
+        internal ILibraryContainer libraryContainer;
 
         // This is to identify whether the PerformShutdownSequenceOnViewModel() method has been
         // called on the view model and the process is not cancelled
@@ -519,74 +519,7 @@ namespace Dynamo.Controls
 
 #else
 
-            try
-            {
-                var geometryRoot = new ItemData();
-                geometryRoot.text = "Geometry";
-                geometryRoot.iconName = "Geometry";
-                geometryRoot.itemType = "category";
-
-                var rootNode = new ItemData();
-                rootNode.childItems.Add(geometryRoot);
-
-                var dir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-                var assembly = Assembly.LoadFrom(Path.Combine(dir, "HostedContents.dll"));
-                var result = assembly.CreateInstance("Dynamo.HostedContents.LibraryContainer");
-                sidebarGrid.Children.Add(result as UserControl);
-
-                ItemData classItemData = null;
-                foreach (var entry in dynamoViewModel.Model.SearchModel.SearchEntries)
-                {
-                    if (!entry.IconName.StartsWith("Autodesk.DesignScript.Geometry.")) continue;
-
-                    var parts = entry.IconName.Split('.');
-                    if ((classItemData == null) || !classItemData.text.Equals(parts[3]))
-                    {
-                        classItemData = new ItemData();
-                        classItemData.text = parts[3];
-                        classItemData.iconName = parts[3];
-                        classItemData.itemType = "none";
-                        geometryRoot.childItems.Add(classItemData);
-                    }
-
-                    var itemType = "none";
-                    var element = entry as NodeSearchElement;
-                    if (element != null)
-                    {
-                        switch (element.Group)
-                        {
-                            case SearchElementGroup.Create: itemType = "creation"; break;
-                            case SearchElementGroup.Action: itemType = "action"; break;
-                            case SearchElementGroup.Query: itemType = "query"; break;
-                        }
-                    }
-
-                    var methodItemData = new ItemData();
-                    methodItemData.text = parts[4];
-                    methodItemData.iconName = entry.IconName;
-                    methodItemData.itemType = itemType;
-                    classItemData.childItems.Add(methodItemData);
-                }
-
-                var outputStream = new MemoryStream();
-                var serializer = new DataContractJsonSerializer(typeof(ItemData));
-                serializer.WriteObject(outputStream, rootNode);
-                outputStream.Position = 0;
-                var inputStream = new StreamReader(outputStream);
-
-                libraryContainer = result as ILibraryContainer;
-                libraryContainer.SetLoadedTypesJson(inputStream.ReadToEnd());
-                libraryContainer.WebBrowserLoaded += (object senderObject, EventArgs eventArgs) =>
-                {
-                    // Does nothing for now.
-                };
-                libraryContainer.SetDynamoViewModel(dynamoViewModel);
-
-                SetLoadedTypesRaw(libraryContainer, dynamoViewModel.Model.SearchModel.SearchEntries);
-            }
-            catch (Exception ex)
-            {
-            }
+            var libraryContainerViewModel = new LibraryContainerViewModel(dynamoViewModel, this);
 
 #endif
 
@@ -679,54 +612,9 @@ namespace Dynamo.Controls
             }
         }
 
-        private void SetLoadedTypesRaw(ILibraryContainer libraryContainer, IEnumerable<NodeSearchElement> searchEntries)
-        {
-            var builder = new System.Text.StringBuilder();
-            builder.Append(@"{ ""loadedTypes"": [");
-
-            bool firstIteration = true;
-            foreach (var entry in searchEntries)
-            {
-                if (firstIteration)
-                    firstIteration = false;
-                else
-                    builder.Append(",");
-
-                var itemType = "";
-                switch (entry.Group)
-                {
-                    case SearchElementGroup.Create: itemType = "creation"; break;
-                    case SearchElementGroup.Action: itemType = "action"; break;
-                    case SearchElementGroup.Query: itemType = "query"; break;
-                }
-
-                builder.Append("{");
-                builder.AppendFormat(" \"fullyQualifiedName\": \"{0}\", ", entry.FullyQualifiedName);
-                builder.AppendFormat(" \"iconName\": \"{0}\", ", entry.IconName);
-                builder.AppendFormat(" \"creationName\": \"{0}\", ", entry.CreationName);
-                builder.AppendFormat(" \"itemType\": \"{0}\" ", itemType);
-                builder.Append("}");
-            }
-
-            builder.Append(@"] }");
-            WriteLibraryToFile(builder);
-            libraryContainer.SetLoadedTypesRaw(builder.ToString());
-        }
-
-        private void WriteLibraryToFile(System.Text.StringBuilder builder)
-        {
-            // Write the unformatted JSON directly in the docs folder in HostedContents
-            string path = System.IO.Directory.GetCurrentDirectory() + "\\..\\..\\..\\src\\HostedContents\\docs\\";
-            using (StreamWriter sw = new StreamWriter(path + "RawTypeData.json"))
-            {
-                sw.WriteLine(builder.ToString());
-            }
-        }
-
         /// <summary>
         /// Close Popup when the Dynamo window is not in the foreground.
         /// </summary>
-
         private void HidePopupWhenWindowDeactivated()
         {
             var workspace = this.ChildOfType<WorkspaceView>();
